+++ CHANGELOG.rst
@@ -2,11 +2,14 @@
 Changelog
 =========
 
+Development version
+====================
+* Improves OPPS generic views queryset to return Containers
+
 0.2.8
 =======
 * Post.related_posts is now DEPRECATED and will removed soon.
     1. Now you can use related_containers instead of related_posts and all Containers have this feature.
-
 
 0.2.7.4
 =======
@@ -24,7 +27,7 @@ Changelog
 
 * Fix bug Opps API, remove page fields on query string
 
-   
+
 0.2.7.1
 =======